Make the code more modular
Create File updater process
Create Documentation module
